add multi select for tables in data sources settings (#278)
* reset data state in useQuery before fetching new data

* remove debounce when fetching tables for airtable base

Input for Airtable base is select, so we don't need debouncing as the
changes are discreet.

* cast values to string in get_block_fallback_content inline with get_remote_value

* add label prop for FieldSelection to allow changing the input label

* allow selecting multiple tables when adding airtable data source

Allows selecting multiple tables from a base when adding Airtable data
source. This removes the ability to select the fields from the table. It
now automatically adds all the fields from the selected tables which are
supported by the plugin.

* add multi select for sheets in google sheets settings

* fetch all sheet fields for spreadsheet at once