add hook to signal plugin loaded (#436)
* add remote_data_blocks_loaded hook to signal plugin is loaded

There would be plugins which build on top of remote data blocks and would hence need to know when this plugin is loaded. The remote_data_blocks_loaded hook signals that remote data blocks is loaded and plugins can safely it.

* avoid loading multiple versions of the plugin

Add a check to prevent multiple versions of the plugin from being loaded by checking if the `REMOTE_DATA_BLOCKS__LOADED` constant is already defined. If it is, it means another instance of the plugin is already loaded and we should not load another version.